adda distance to views, bug z decimal z przecinkiem

+++ Dziennik_v0.1/Core/Models/CardioModel.cs
@@ -2,6 +2,7 @@
 using System;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
 using System.Web;
 
@@ -27,7 +28,7 @@ namespace Dziennik_v0._1.Core.Models
 
         [Required]
         [Display(Name = "Dystans:")]
-        public double Distance { get; set; }
+        public decimal Distance { get; set; }
 
         [Required]
         [Display(Name = "Rodzaj treningu:")]
